requirements.txt
openai==1.30.1
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
retrying==1.3.4
azure-common==1.1.28
azure-core==1.29.4
python-dotenv==1.0.1
streamlit==1.34.0
